+++ src/main/java/org/privacyidea/authenticator/Configuration.java
@@ -22,10 +22,8 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import org.jboss.logging.Logger;
 
 import static org.privacyidea.authenticator.Const.CONFIG_CUSTOM_HEADERS;
-import static org.privacyidea.authenticator.Const.CONFIG_DEFAULT_MESSAGE;
 import static org.privacyidea.authenticator.Const.CONFIG_ENABLE_LOG;
 import static org.privacyidea.authenticator.Const.CONFIG_EXCLUDED_GROUPS;
 import static org.privacyidea.authenticator.Const.CONFIG_FORWARDED_HEADERS;
@@ -34,7 +32,6 @@ import static org.privacyidea.authenticator.Const.CONFIG_INCLUDED_GROUPS;
 import static org.privacyidea.authenticator.Const.CONFIG_OTP_LENGTH;
 import static org.privacyidea.authenticator.Const.CONFIG_POLL_IN_BROWSER;
 import static org.privacyidea.authenticator.Const.CONFIG_POLL_IN_BROWSER_URL;
-import static org.privacyidea.authenticator.Const.CONFIG_PUSH_INTERVAL;
 import static org.privacyidea.authenticator.Const.CONFIG_REALM;
 import static org.privacyidea.authenticator.Const.CONFIG_SEND_PASSWORD;
 import static org.privacyidea.authenticator.Const.CONFIG_SEND_STATIC_PASS;
@@ -45,8 +42,7 @@ import static org.privacyidea.authenticator.Const.CONFIG_SERVICE_REALM;
 import static org.privacyidea.authenticator.Const.CONFIG_STATIC_PASS;
 import static org.privacyidea.authenticator.Const.CONFIG_TRIGGER_CHALLENGE;
 import static org.privacyidea.authenticator.Const.CONFIG_VERIFY_SSL;
-import static org.privacyidea.authenticator.Const.DEFAULT_POLLING_ARRAY;
-import static org.privacyidea.authenticator.Const.DEFAULT_POLLING_INTERVAL;
+import static org.privacyidea.authenticator.Const.POLLING_INTERVALS;
 import static org.privacyidea.authenticator.Const.TRUE;
 
 
@@ -72,11 +68,8 @@ public class Configuration
     private final String pollInBrowserUrl;
     private final List<Integer> pollingInterval = new ArrayList<>();
     private final int configHash;
-    private final String defaultOTPMessage;
     private final Map<String, String> customHeaders = new HashMap<>();
-
-    private final Logger logger = Logger.getLogger(PrivacyIDEAAuthenticator.class);
-
+    private final int httpTimeoutMs;
 
     public Configuration(Map<String, String> configMap)
     {
@@ -90,7 +83,6 @@ public class Configuration
         this.serviceAccountRealm = configMap.get(CONFIG_SERVICE_REALM) == null ? "" : configMap.get(CONFIG_SERVICE_REALM);
         this.staticPass = configMap.get(CONFIG_STATIC_PASS) == null ? "" : configMap.get(CONFIG_STATIC_PASS);
         this.forwardClientIP = configMap.get(CONFIG_FORWARD_CLIENT_IP) != null && configMap.get(CONFIG_FORWARD_CLIENT_IP).equals(TRUE);
-        this.defaultOTPMessage = configMap.get(CONFIG_DEFAULT_MESSAGE) == null ? "" : configMap.get(CONFIG_DEFAULT_MESSAGE);
         this.otpLength = configMap.get(CONFIG_OTP_LENGTH) == null ? "" : configMap.get(CONFIG_OTP_LENGTH);
         this.pollInBrowser = (configMap.get(CONFIG_POLL_IN_BROWSER) != null && configMap.get(CONFIG_POLL_IN_BROWSER).equals(TRUE));
         this.pollInBrowserUrl = configMap.get(CONFIG_POLL_IN_BROWSER_URL) == null ? "" : configMap.get(CONFIG_POLL_IN_BROWSER_URL);
@@ -116,31 +108,8 @@ public class Configuration
             this.forwardedHeaders.addAll(Arrays.asList(forwardedHeadersStr.split(",")));
         }
 
-        // Set intervals to either default or configured values
-        String s = configMap.get(CONFIG_PUSH_INTERVAL);
-        if (s != null)
-        {
-            List<String> strPollingIntervals = Arrays.asList(s.split(","));
-            if (!strPollingIntervals.isEmpty())
-            {
-                this.pollingInterval.clear();
-                for (String str : strPollingIntervals)
-                {
-                    try
-                    {
-                        this.pollingInterval.add(Integer.parseInt(str));
-                    }
-                    catch (NumberFormatException e)
-                    {
-                        this.pollingInterval.add(DEFAULT_POLLING_INTERVAL);
-                    }
-                }
-            }
-        }
-        else
-        {
-            this.pollingInterval.addAll(DEFAULT_POLLING_ARRAY);
-        }
+        this.pollingInterval.addAll(POLLING_INTERVALS);
+
         String entry = configMap.get(CONFIG_CUSTOM_HEADERS);
         String[] entries = entry.split("##");
         for (String e : entries)
@@ -150,12 +119,9 @@ public class Configuration
             {
                 customHeaders.put(keyValue[0], keyValue[1]);
             }
-            else
-            {
-                logger.error("Invalid custom header entry: " + e);
-            }
         }
-        logger.error("Header: " + customHeaders);
+
+        this.httpTimeoutMs = Integer.parseInt(configMap.getOrDefault("httpTimeoutMs", "10000"));
     }
 
     int configHash()
@@ -258,13 +224,13 @@ public class Configuration
         return doSendPassword;
     }
 
-    String defaultOTPMessage()
-    {
-        return defaultOTPMessage;
-    }
-
     Map<String, String> customHeaders()
     {
         return customHeaders;
+    }
+
+    public int httpTimeoutMs()
+    {
+        return httpTimeoutMs;
     }
 }