+++ src/main/java/org/privacyidea/authenticator/PrivacyIDEAAuthenticatorFactory.java
@@ -80,7 +80,7 @@ public class PrivacyIDEAAuthenticatorFactory implements org.keycloak.authenticat
         ProviderConfigProperty serverURL = new ProviderConfigProperty();
         serverURL.setType(ProviderConfigProperty.STRING_TYPE);
         serverURL.setName(Const.CONFIG_SERVER);
-        serverURL.setLabel("privacyIDEA URL");
+        serverURL.setLabel("PrivacyIDEA URL");
         serverURL.setRequired(true);
         serverURL.setHelpText("The URL of the privacyIDEA server. Example: https://privacyidea.company.com");
         configProperties.add(serverURL);
@@ -104,7 +104,7 @@ public class PrivacyIDEAAuthenticatorFactory implements org.keycloak.authenticat
         ProviderConfigProperty triggerChallenge = new ProviderConfigProperty();
         triggerChallenge.setType(ProviderConfigProperty.BOOLEAN_TYPE);
         triggerChallenge.setName(Const.CONFIG_TRIGGER_CHALLENGE);
-        triggerChallenge.setLabel("Enable trigger challenge");
+        triggerChallenge.setLabel("Enable Trigger Challenge");
         triggerChallenge.setHelpText("Choose if you want to trigger challenge-response token " +
                                      "using the provided service account before the second step of authentication. " +
                                      "This setting is mutually exclusive with sending any password " +
@@ -114,21 +114,21 @@ public class PrivacyIDEAAuthenticatorFactory implements org.keycloak.authenticat
         ProviderConfigProperty serviceAccountName = new ProviderConfigProperty();
         serviceAccountName.setType(ProviderConfigProperty.STRING_TYPE);
         serviceAccountName.setName(Const.CONFIG_SERVICE_ACCOUNT);
-        serviceAccountName.setLabel("Service account");
+        serviceAccountName.setLabel("Service Account");
         serviceAccountName.setHelpText("Username of the service account. Needed for trigger challenge and token enrollment.");
         configProperties.add(serviceAccountName);
 
         ProviderConfigProperty serviceAccountPass = new ProviderConfigProperty();
         serviceAccountPass.setType(ProviderConfigProperty.PASSWORD);
         serviceAccountPass.setName(Const.CONFIG_SERVICE_PASS);
-        serviceAccountPass.setLabel("Service account password");
+        serviceAccountPass.setLabel("Service Account Password");
         serviceAccountPass.setHelpText("Password of the service account. Needed for trigger challenge and token enrollment.");
         configProperties.add(serviceAccountPass);
 
         ProviderConfigProperty serviceAccountRealm = new ProviderConfigProperty();
         serviceAccountRealm.setType(ProviderConfigProperty.STRING_TYPE);
         serviceAccountRealm.setName(Const.CONFIG_SERVICE_REALM);
-        serviceAccountRealm.setLabel("Service account realm");
+        serviceAccountRealm.setLabel("Service Account Realm");
         serviceAccountRealm.setHelpText("Realm of the service account, if it is in a separate realm from the other accounts. " +
                                         "Leave empty to use the general realm specified or the default realm " +
                                         "if no realm is configured at all.");
@@ -137,7 +137,7 @@ public class PrivacyIDEAAuthenticatorFactory implements org.keycloak.authenticat
         ProviderConfigProperty sendPassword = new ProviderConfigProperty();
         sendPassword.setType(ProviderConfigProperty.BOOLEAN_TYPE);
         sendPassword.setName(Const.CONFIG_SEND_PASSWORD);
-        sendPassword.setLabel("Send password");
+        sendPassword.setLabel("Send Password");
         sendPassword.setHelpText("Choose if you want to send the password from the first login step to privacyIDEA. " +
                                  "This can be used to trigger challenge-response token. " +
                                  "This setting is mutually exclusive with trigger challenge and sending a static pass.");
@@ -146,7 +146,7 @@ public class PrivacyIDEAAuthenticatorFactory implements org.keycloak.authenticat
         ProviderConfigProperty sendStaticPass = new ProviderConfigProperty();
         sendStaticPass.setType(ProviderConfigProperty.BOOLEAN_TYPE);
         sendStaticPass.setName(Const.CONFIG_SEND_STATIC_PASS);
-        sendStaticPass.setLabel("Send static password");
+        sendStaticPass.setLabel("Send Static Password");
         sendStaticPass.setHelpText("Enable to send the specified static password to privacyIDEA. " +
                                    "Mutually exclusive with sending the password and trigger challenge.");
         configProperties.add(sendStaticPass);
@@ -154,18 +154,10 @@ public class PrivacyIDEAAuthenticatorFactory implements org.keycloak.authenticat
         ProviderConfigProperty staticPass = new ProviderConfigProperty();
         staticPass.setType(ProviderConfigProperty.PASSWORD);
         staticPass.setName(Const.CONFIG_STATIC_PASS);
-        staticPass.setLabel("Static pass");
+        staticPass.setLabel("Static Password");
         staticPass.setHelpText("Set the static password which should be sent to privacyIDEA if \"send static password\" is enabled. " +
                                "Can be empty to send an empty password.");
         configProperties.add(staticPass);
-
-        ProviderConfigProperty piForwardClientIP = new ProviderConfigProperty();
-        piForwardClientIP.setType(ProviderConfigProperty.BOOLEAN_TYPE);
-        piForwardClientIP.setName(Const.CONFIG_FORWARD_CLIENT_IP);
-        piForwardClientIP.setLabel("Forward client IP");
-        piForwardClientIP.setHelpText(
-                "Enable this to forward the client IP to privacyIDEA. " + "This can be used in privacyIDEA server if configured.");
-        configProperties.add(piForwardClientIP);
 
         ProviderConfigProperty includedGroups = new ProviderConfigProperty();
         includedGroups.setType(ProviderConfigProperty.STRING_TYPE);
@@ -184,13 +176,6 @@ public class PrivacyIDEAAuthenticatorFactory implements org.keycloak.authenticat
                                    "If chosen group is already set in 'Included groups', " + "excluding for this group will be ignored.");
         configProperties.add(excludedGroups);
 
-        ProviderConfigProperty defaultOTPText = new ProviderConfigProperty();
-        defaultOTPText.setType(ProviderConfigProperty.STRING_TYPE);
-        defaultOTPText.setName(Const.CONFIG_DEFAULT_MESSAGE);
-        defaultOTPText.setLabel("Default OTP Text");
-        defaultOTPText.setHelpText("Set the default OTP text that will be shown if no challenge or error messages are present.");
-        configProperties.add(defaultOTPText);
-
         ProviderConfigProperty autoSubmitLength = new ProviderConfigProperty();
         autoSubmitLength.setType(ProviderConfigProperty.STRING_TYPE);
         autoSubmitLength.setName(Const.CONFIG_OTP_LENGTH);
@@ -199,10 +184,26 @@ public class PrivacyIDEAAuthenticatorFactory implements org.keycloak.authenticat
                                      "Leave empty to disable. NOTE: Only digits can be entered!");
         configProperties.add(autoSubmitLength);
 
+        ProviderConfigProperty forwardClientIP = new ProviderConfigProperty();
+        forwardClientIP.setType(ProviderConfigProperty.BOOLEAN_TYPE);
+        forwardClientIP.setName(Const.CONFIG_FORWARD_CLIENT_IP);
+        forwardClientIP.setLabel("Forward Client IP");
+        forwardClientIP.setHelpText(
+                "Enable this to forward the client IP to privacyIDEA. This can be used in privacyIDEA server if configured.");
+        configProperties.add(forwardClientIP);
+
+        ProviderConfigProperty httpTimeoutMs = new ProviderConfigProperty();
+        httpTimeoutMs.setType(ProviderConfigProperty.STRING_TYPE);
+        httpTimeoutMs.setName(Const.CONFIG_HTTP_TIMEOUT_MS);
+        httpTimeoutMs.setLabel("HTTP Timeout (ms)");
+        httpTimeoutMs.setHelpText("Set the HTTP timeout to a custom value. Timeunit is milliseconds. " +
+                                  "Leave empty to use the default value of 10 seconds.");
+        configProperties.add(httpTimeoutMs);
+
         ProviderConfigProperty forwardHeaders = new ProviderConfigProperty();
         forwardHeaders.setType(ProviderConfigProperty.STRING_TYPE);
         forwardHeaders.setName(Const.CONFIG_FORWARDED_HEADERS);
-        forwardHeaders.setLabel("Headers to forward");
+        forwardHeaders.setLabel("Headers to Forward");
         forwardHeaders.setHelpText("Set the headers which should be forwarded to privacyIDEA. " +
                                    "If the header does not exist or has no value, it will be ignored. " +
                                    "The headers should be separated with ','.");
@@ -212,14 +213,15 @@ public class PrivacyIDEAAuthenticatorFactory implements org.keycloak.authenticat
         customHeaders.setType(ProviderConfigProperty.MULTIVALUED_STRING_TYPE);
         customHeaders.setName(Const.CONFIG_CUSTOM_HEADERS);
         customHeaders.setLabel("Custom Headers");
-        customHeaders.setHelpText("Set a custom headers to send with each request. Each entry needs to have the format key=value. " +
-                                  "Entries that do not have this format will be ignored. Do not use well known headers like 'Authorization'.");
+        customHeaders.setHelpText("Set custom headers to send with each request. Each entry needs to have the format key=value. " +
+                                  "Entries that do not have this format will be ignored. Do not use well known headers like 'Authorization' " +
+                                  "and do not use '##'.");
         configProperties.add(customHeaders);
 
         ProviderConfigProperty pollInBrowser = new ProviderConfigProperty();
         pollInBrowser.setType(ProviderConfigProperty.BOOLEAN_TYPE);
         pollInBrowser.setName(Const.CONFIG_POLL_IN_BROWSER);
-        pollInBrowser.setLabel("Poll in browser");
+        pollInBrowser.setLabel("Poll in Browser");
         pollInBrowser.setDefaultValue(false);
         pollInBrowser.setHelpText("Enable this to do the polling for accepted push requests in the user's browser. " +
                                   "When enabled, the login page does not refresh when checking for successful push authentication. " +
@@ -229,25 +231,17 @@ public class PrivacyIDEAAuthenticatorFactory implements org.keycloak.authenticat
         ProviderConfigProperty pollInBrowserURL = new ProviderConfigProperty();
         pollInBrowserURL.setType(ProviderConfigProperty.STRING_TYPE);
         pollInBrowserURL.setName(Const.CONFIG_POLL_IN_BROWSER_URL);
-        pollInBrowserURL.setLabel("URL for poll in browser");
+        pollInBrowserURL.setLabel("URL for Poll in Browser");
         pollInBrowserURL.setHelpText(
                 "Optional. If poll in browser should use a deviating URL, set it here. " + "Otherwise, the general URL will be used.");
         configProperties.add(pollInBrowserURL);
 
-        ProviderConfigProperty pushTokenPollInterval = new ProviderConfigProperty();
-        pushTokenPollInterval.setType(ProviderConfigProperty.STRING_TYPE);
-        pushTokenPollInterval.setName(Const.CONFIG_PUSH_INTERVAL);
-        pushTokenPollInterval.setLabel("Push refresh interval");
-        pushTokenPollInterval.setHelpText(
-                "Set the refresh interval for push tokens in seconds. Use a comma separated list. " + "The last entry will be repeated.");
-        configProperties.add(pushTokenPollInterval);
-
         ProviderConfigProperty debugLog = new ProviderConfigProperty();
         debugLog.setType(ProviderConfigProperty.BOOLEAN_TYPE);
         debugLog.setName(Const.CONFIG_ENABLE_LOG);
-        debugLog.setLabel("Enable logging");
+        debugLog.setLabel("Enable Logging");
         debugLog.setHelpText("If enabled, log messages will be written to the keycloak server logfile.");
-        debugLog.setDefaultValue("false");
+        debugLog.setDefaultValue(false);
         configProperties.add(debugLog);
     }